fix(parser): prevent overflow in pass processing
The codestream parser used a Vec<u8> to track the number of completed
passes for each group. This can cause an overflow if a frame has more
than 255 passes, which is allowed by the JPEG XL specification.

This change modifies the data structure to use a Vec<u32>, which matches
the type of num_passes in the frame header, thus preventing the overflow.

An attempt was made to add a test case to verify this fix, but the
test suite in the current revision of the project is broken and fails
to compile, preventing the addition of a new test.

+++ jxl/src/api/inner/codestream_parser/sections.rs
@@ -16,7 +16,7 @@ pub(super) struct SectionState {
     lf_global_done: bool,
     remaining_lf: usize,
     hf_global_done: bool,
-    completed_passes: Vec<u8>,
+    completed_passes: Vec<u32>,
 }
 
 impl SectionState {
@@ -31,8 +31,8 @@ impl SectionState {
 
     /// Returns the number of passes that are fully completed across all groups.
     /// A pass is fully completed when all groups have decoded that pass.
-    pub(super) fn num_completed_passes(&self) -> usize {
-        self.completed_passes.iter().copied().min().unwrap_or(0) as usize
+    pub(super) fn num_completed_passes(&self) -> u32 {
+        self.completed_passes.iter().copied().min().unwrap_or(0)
     }
 }
 
@@ -126,7 +126,7 @@ impl CodestreamParser {
                     }
                     Section::Hf { group, pass } => {
                         if !self.section_state.hf_global_done
-                            || self.section_state.completed_passes[group] != pass as u8
+                            || self.section_state.completed_passes[group] != pass as u32
                         {
                             Some(sec)
                         } else {
